fix warning

+++ libsandbox.c
@@ -54,7 +54,7 @@ uint8_t *patch_cursor = NULL;
 
 uint64_t get_sandbox_start(void)
 {
-	return  &patch_sandbox_start;
+	return  (uint64_t)&patch_sandbox_start;
 }
 
 uint64_t get_sandbox_end(void)
